+++ src/bindings.rs
@@ -1296,7 +1296,7 @@ fn bindgen_test_layout__rknn_output_extend() {
     );
 }
 pub type rknn_output_extend = _rknn_output_extend;
-extern "C" {
+unsafe extern "C" {
     pub fn rknn_init(
         context: *mut rknn_context,
         model: *mut ::std::os::raw::c_void,
@@ -1304,73 +1304,49 @@ extern "C" {
         flag: u32,
         extend: *mut rknn_init_extend,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn rknn_dup_context(
+    pub unsafe fn rknn_dup_context(
         context_in: *mut rknn_context,
         context_out: *mut rknn_context,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_destroy(context: rknn_context) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_query(
         context: rknn_context,
         cmd: rknn_query_cmd,
         info: *mut ::std::os::raw::c_void,
         size: u32,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_inputs_set(
         context: rknn_context,
         n_inputs: u32,
         inputs: *mut rknn_input,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn rknn_set_batch_core_num(
+    pub unsafe fn rknn_set_batch_core_num(
         context: rknn_context,
         core_num: ::std::os::raw::c_int,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_set_core_mask(
         context: rknn_context,
         core_mask: rknn_core_mask,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_run(context: rknn_context, extend: *mut rknn_run_extend) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_wait(context: rknn_context, extend: *mut rknn_run_extend) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_outputs_get(
         context: rknn_context,
         n_outputs: u32,
         outputs: *mut rknn_output,
         extend: *mut rknn_output_extend,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_outputs_release(
         context: rknn_context,
         n_ouputs: u32,
         outputs: *mut rknn_output,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_create_mem_from_phys(
         ctx: rknn_context,
         phys_addr: u64,
         virt_addr: *mut ::std::os::raw::c_void,
         size: u32,
     ) -> *mut rknn_tensor_mem;
-}
-extern "C" {
     pub fn rknn_create_mem_from_fd(
         ctx: rknn_context,
         fd: i32,
@@ -1378,57 +1354,37 @@ extern "C" {
         size: u32,
         offset: i32,
     ) -> *mut rknn_tensor_mem;
-}
-extern "C" {
     pub fn rknn_create_mem_from_mb_blk(
         ctx: rknn_context,
         mb_blk: *mut ::std::os::raw::c_void,
         offset: i32,
     ) -> *mut rknn_tensor_mem;
-}
-extern "C" {
     pub fn rknn_create_mem(ctx: rknn_context, size: u32) -> *mut rknn_tensor_mem;
-}
-extern "C" {
     pub fn rknn_create_mem2(ctx: rknn_context, size: u64, alloc_flags: u64)
         -> *mut rknn_tensor_mem;
-}
-extern "C" {
     pub fn rknn_destroy_mem(ctx: rknn_context, mem: *mut rknn_tensor_mem) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_set_weight_mem(
         ctx: rknn_context,
         mem: *mut rknn_tensor_mem,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_set_internal_mem(
         ctx: rknn_context,
         mem: *mut rknn_tensor_mem,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_set_io_mem(
         ctx: rknn_context,
         mem: *mut rknn_tensor_mem,
         attr: *mut rknn_tensor_attr,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_set_input_shape(
         ctx: rknn_context,
         attr: *mut rknn_tensor_attr,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_set_input_shapes(
         ctx: rknn_context,
         n_inputs: u32,
         attr: *mut rknn_tensor_attr,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_mem_sync(
         context: rknn_context,
         mem: *mut rknn_tensor_mem,
@@ -1855,14 +1811,12 @@ fn bindgen_test_layout__rknn_custom_op() {
 pub type rknn_custom_op = _rknn_custom_op;
 #[doc = " dlopen custom op with so required this function"]
 pub type get_custom_op_func = ::std::option::Option<unsafe extern "C" fn() -> *mut rknn_custom_op>;
-extern "C" {
+unsafe extern "C" {
     pub fn rknn_register_custom_ops(
         ctx: rknn_context,
         op: *mut rknn_custom_op,
         custom_op_num: u32,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_custom_op_get_op_attr(
         op_ctx: *mut rknn_custom_op_context,
         attr_name: *const ::std::os::raw::c_char,
@@ -2269,14 +2223,12 @@ fn bindgen_test_layout_rknn_matmul_info_t() {
     );
 }
 pub type rknn_matmul_info = rknn_matmul_info_t;
-extern "C" {
+unsafe extern "C" {
     pub fn rknn_matmul_create(
         ctx: *mut rknn_matmul_ctx,
         info: *mut rknn_matmul_info,
         io_attr: *mut rknn_matmul_io_attr,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_matmul_create_dyn_shape(
         ctx: *mut rknn_matmul_ctx,
         info: *mut rknn_matmul_info,
@@ -2284,46 +2236,30 @@ extern "C" {
         dynamic_shapes: *mut rknn_matmul_shape,
         io_attrs: *mut rknn_matmul_io_attr,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_matmul_set_io_mem(
         ctx: rknn_matmul_ctx,
         mem: *mut rknn_tensor_mem,
         attr: *mut rknn_matmul_tensor_attr,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_matmul_set_core_mask(
         context: rknn_matmul_ctx,
         core_mask: rknn_core_mask,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_matmul_set_quant_params(
         context: rknn_matmul_ctx,
         params: *mut rknn_quant_params,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_matmul_get_quant_params(
         ctx: rknn_matmul_ctx,
         params: *mut rknn_quant_params,
         scale: *mut f32,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_matmul_set_dynamic_shape(
         ctx: rknn_matmul_ctx,
         shape: *mut rknn_matmul_shape,
     ) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_matmul_run(ctx: rknn_matmul_ctx) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_matmul_destroy(ctx: rknn_matmul_ctx) -> ::std::os::raw::c_int;
-}
-extern "C" {
     pub fn rknn_B_normal_layout_to_native_layout(
         B_input: *mut ::std::os::raw::c_void,
         B_output: *mut ::std::os::raw::c_void,